+++ lpmalgos/src/implementation/geodesics.h
@@ -9,10 +9,10 @@ double angular_distance(const Location &A,
                         const Location &B,
                         const Location &C);
 
-std::vector<int> find_clusters(const Locations &locs,
-                               const Ellipsoid &ani,
-                               double r_tol, double angular_tol,
-                               double support_threshold,
-                               int min_support_size);
+std::vector<int16_t> find_clusters(const Locations &locs,
+                                   const Ellipsoid &ani,
+                                   double r_tol, double angular_tol,
+                                   double support_threshold,
+                                   size_t min_support_size);
 
 } // namespace lpmalgos
